refund-agent-mcp/pyproject.toml
[project]
name = "refund-agent-mcp"
version = "0.1.0"
description = "Refund Agent leveraging the Portia SDK MCP integration with Stripe"
authors = [
    {name = "Portia AI",email = "hello@portialabs.ai"}
]
readme = "README.md"

requires-python = ">=3.11,<4.0"

dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "portia-sdk-python[all] (>=0.3.0,<0.4.0)",
    "stripe (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false